Zero-shot_experiments/moe_requirement.txt
accelerate==0.21.0
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
anyio==4.3.0
attrs==23.2.0
bitsandbytes==0.41.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cmake==3.29.3
contourpy==1.2.1
cycler==0.12.1
deepspeed==0.14.2
dnspython==2.6.1
einops==0.6.1
einops-exts==0.0.4
email_validator==2.1.1
et-xmlfile==1.1.0
fastapi==0.111.0
fastapi-cli==0.0.3
ffmpy==0.3.2
filelock==3.14.0
flash-attn==2.5.8
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.3.1
gradio==3.37.0
gradio_client==0.7.0
h11==0.14.0
hjson==3.1.0
httpcore==0.17.3
httptools==0.6.1
httpx==0.24.0
huggingface-hub==0.23.0
idna==3.7
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
linkify-it-py==2.0.3
lit==18.1.4
markdown-it-py==2.2.0
markdown2==2.4.13
MarkupSafe==2.1.5
matplotlib==3.8.4
mdit-py-plugins==0.3.3
mdurl==0.1.2
-e git+https://github.com/PKU-YuanGroup/MoE-LLaVA@454710663485a6c91f874b40d782fb2e936f7309#egg=moellava
mpi4py_mpich==3.1.5
mpmath==1.3.0
multidict==6.0.5
networkx==3.3
ninja==1.11.1.1
numpy==1.26.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
openai==0.28.0
openpyxl==3.1.2
orjson==3.10.3
packaging==24.0
pandas==2.2.2
peft==0.4.0
pillow==10.3.0
protobuf==4.24.4
psutil==5.9.8
py-cpuinfo==9.0.0
pydantic==1.10.15
pydub==0.25.1
Pygments==2.18.0
pynvml==11.5.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
regex==2024.5.10
requests==2.31.0
rich==13.7.1
rpds-py==0.18.1
safetensors==0.4.3
scikit-learn==1.2.2
scipy==1.13.0
semantic-version==2.10.0
sentencepiece==0.1.99
shellingham==1.5.4
shortuuid==1.0.13
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
svgwrite==1.4.3
sympy==1.12
threadpoolctl==3.5.0
tiktoken==0.5.2
timm==0.6.13
tokenizers==0.15.1
toolz==0.12.1
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.4
transformers==4.37.0
triton==2.0.0
typer==0.12.3
typing_extensions==4.11.0
tzdata==2024.1
uc-micro-py==1.0.3
ujson==5.10.0
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
wavedrom==2.0.3.post3
websockets==11.0.3
yarl==1.9.4
